requirements.txt
mysql-connector-python==9.2.0
shapely==2.0.7
python-dotenv==1.0.1
requests==2.32.3 
pyproj==3.7.1
pika==1.3.2
geopandas==1.0.1
rasterio==1.4.3